# Windows: base implementation on `GetTimeZoneInformationForYear`