--- a/src/datetime/tests.rs
+++ b/src/datetime/tests.rs
@@ -1500,41 +1500,6 @@ fn test_core_duration_max() {
 }
 
 #[test]
-#[cfg(all(target_os = "windows", feature = "clock"))]
-fn test_from_naive_date_time_windows() {
-    let min_year = NaiveDate::from_ymd_opt(1601, 1, 3).unwrap().and_hms_opt(0, 0, 0).unwrap();
-
-    let max_year = NaiveDate::from_ymd_opt(30827, 12, 29).unwrap().and_hms_opt(23, 59, 59).unwrap();
-
-    let too_low_year =
-        NaiveDate::from_ymd_opt(1600, 12, 29).unwrap().and_hms_opt(23, 59, 59).unwrap();
-
-    let too_high_year = NaiveDate::from_ymd_opt(30829, 1, 3).unwrap().and_hms_opt(0, 0, 0).unwrap();
-
-    let _ = Local.from_utc_datetime(&min_year);
-    let _ = Local.from_utc_datetime(&max_year);
-
-    let _ = Local.from_local_datetime(&min_year);
-    let _ = Local.from_local_datetime(&max_year);
-
-    let local_too_low = Local.from_local_datetime(&too_low_year);
-    let local_too_high = Local.from_local_datetime(&too_high_year);
-
-    assert_eq!(local_too_low, LocalResult::None);
-    assert_eq!(local_too_high, LocalResult::None);
-
-    let err = std::panic::catch_unwind(|| {
-        Local.from_utc_datetime(&too_low_year);
-    });
-    assert!(err.is_err());
-
-    let err = std::panic::catch_unwind(|| {
-        Local.from_utc_datetime(&too_high_year);
-    });
-    assert!(err.is_err());
-}
-
-#[test]
 #[cfg(feature = "clock")]
 fn test_datetime_local_from_preserves_offset() {
     let naivedatetime = NaiveDate::from_ymd_opt(2023, 1, 1).unwrap().and_hms_opt(0, 0, 0).unwrap();