--- a/src/offset/local/mod.rs
+++ b/src/offset/local/mod.rs
@@ -292,8 +292,7 @@ mod tests {
 
     #[test]
     fn verify_correct_offsets_distant_past() {
-        // let distant_past = Local::now() - TimeDelta::days(365 * 100);
-        let distant_past = Local::now() - TimeDelta::days(250 * 31);
+        let distant_past = Local::now() - TimeDelta::days(365 * 500);
         let from_local = Local.from_local_datetime(&distant_past.naive_local()).unwrap();
         let from_utc = Local.from_utc_datetime(&distant_past.naive_utc());
 
@@ -306,7 +305,7 @@ mod tests {
 
     #[test]
     fn verify_correct_offsets_distant_future() {
-        let distant_future = Local::now() + TimeDelta::days(250 * 31);
+        let distant_future = Local::now() + TimeDelta::days(365 * 35000);
         let from_local = Local.from_local_datetime(&distant_future.naive_local()).unwrap();
         let from_utc = Local.from_utc_datetime(&distant_future.naive_utc());