--- a/src/offset/local/win_bindings.txt
+++ b/src/offset/local/win_bindings.txt
@@ -1,6 +1,7 @@
 --out src/offset/local/win_bindings.rs
 --config flatten sys
 --filter
+    Windows.Win32.System.Time.GetTimeZoneInformationForYear
     Windows.Win32.System.Time.SystemTimeToFileTime
     Windows.Win32.System.Time.SystemTimeToTzSpecificLocalTime
     Windows.Win32.System.Time.TzSpecificLocalTimeToSystemTime