… src/offset/local/win_bindings.rs
… src/offset/local/windows.rs