Issues caused by the current use of Windows API's:
- `TzSpecificLocalTimeToSystemTime` doesn't let us choose how to handle ambiguous cases during a DST transition. It always returns the earliest option.
- using `GetLocalTime` with `TzSpecificLocalTimeToSystemTime` can give wrong results during a DST transition  because the issue mentioned above.
- `TzSpecificLocalTimeToSystemTime` always returns a result, also for non-existing datetimes during a DST transition.
- The Windows API restricts years to a range of `1601..30828`. 

The alternative in this PR is to calculate the offset ourselves using the information provided by `GetTimeZoneInformationForYear`.This is similar to what we do on Unix.

The tricky bits to determine the correct offset between DST transitions is copied from the Unix implementation. Initially I wanted to share the code between the two systems, but that didn't make the Unix implementation any clearer.

Part of the old Windows implementation have been moved to a test, so our implementation can be compared to `TzSpecificLocalTimeToSystemTime`. I have set my computer timezone to ca. 20 'interesting' timezones (determined by looking at data of the entries enumerated by [`EnumDynamicTimeZoneInformation`](https://learn.microsoft.com/en-us/windows/win32/api/timezoneapi/nf-timezoneapi-enumdynamictimezoneinformation)) and found no differences.

---

I ended up writing this PR because of a yak shaving. In order to remove some `unwrap`s from my code, I need some progress on https://github.com/chronotope/chrono/issues/716. To show we need a fourth enum variant for `LocalResult`, I started implementing a proof of concept for `DateTime<Local>`. And it turned out the Windows implementation had no place to hook into...

Fixes https://github.com/chronotope/chrono/issues/1150, https://github.com/chronotope/chrono/issues/651.