@esheppa 2023-04-17
It looks like these two PRs should be somewhat orthogonal, but this is also related to https://github.com/chronotope/chrono/pull/750 which is great as by extending this we should be able to extract the name of the timezone as well