@djc 2023-04-17
> The tricky bits to determine the correct offset between DST transitions is copied from the Unix implementation. Initially I wanted to share the code between the two systems, but that didn't make the Unix implementation any clearer.

Can you add that here as a separate commit? It does sound like something we'd want.