@pitdicker 2023-04-17
> Did you see #997 already?

Somewhat, I forgot about it. But the PR before by @nekevss was very useful for feeling my way in the Windows implementation.

> Can you add that here as a separate commit? It does sound like something we'd want.

I can dig up the commit, but it will probably be tomorrow because it needs some cleaning up.