@esheppa 2023-04-17
We could also review and merge this as is, before enhancing to get the timezone name via another PR if that is more suitable @djc?