@pitdicker 2023-04-19
Just noticed issue https://github.com/chronotope/chrono/issues/651. That issue would be fixed by this PR, added a test.