@pitdicker 2023-04-24
I noticed `Utc::now` uses the standard library to get the current date, while for `Local::now` we call a Windows API ourselves. Because on both Unix and Windows we get an UTC timestamp from the OS and then convert it to local time, we may just as well always use `Utc::now` inside `Local::now` and be consistent.

On Windows the standard library uses [`QueryPerformanceCounter`](https://learn.microsoft.com/en-us/windows/win32/api/profileapi/nf-profileapi-queryperformancecounter), which is guaranteed to provide sub-microsecond values. What we used for `Local::now` returned a `SYSTEMTIME` with at most millisecond precision.