@pitdicker 2023-04-24
The benchmark results have improved a bit.
Before:
```
     Running benches\chrono.rs (target\release\deps\chrono-3f8222dc2c00b8c6.exe)
bench_get_local_time    time:   [1.4577 µs 1.5100 µs 1.5519 µs]
```

After:
```
     Running benches\chrono.rs (target\release\deps\chrono-3f8222dc2c00b8c6.exe)
bench_get_local_time    time:   [1.2091 µs 1.2737 µs 1.3304 µs]
                        change: [-17.784% -13.853% -9.6609%] (p = 0.00 < 0.05)
                        Performance has improved.
```