@pitdicker 2023-04-30
@jtmoon79 Thank you for the review again!

I am a slow learning when it comes to adding tests :rofl:. This will take some work.