@pitdicker 2023-05-04
Now that Windows supports for greater year ranges, it might overflow when near the limits of `NaiveDateTime`.
When adding a test for creating a `DateTime<Local>` with `NaiveDateTime::{MIN, MAX}` I hit multiple fun issues.
- It panicked on overflow instead of returning `LocalResult::None`.
- The `Debug` implementation of `DateTime` can panic (not something I want to fix in this PR).
- `FixedOffset` has the same issue.

This can be fixed inside the provided implementation of `TimeZone::from_local_datetime`, which I did.

But `Local` doesn't use that, the code is duplicated.
`Local` also does other things in a roundabout way, creating a `DateTime` just to extract the offset in `offset_from_local_datetime` for example. I would like to clean this up, but that would complicate this PR even more. I'll wait until after this is merged.