@pitdicker 2023-05-05
> `Local` also does other things in a roundabout way, creating a `DateTime` just to extract the offset in `offset_from_local_datetime` for example. I would like to clean this up, but that would complicate this PR even more. I'll wait until after this is merged.

Thought better of this. I have removed the commit this fixes handling overflow when near the limits of `NaiveDateTime`.

I have added a commit to this PR removing all the roundabout stuff in the `local` module, including in `stub.rs`. In my opinion it is quite a bit more readable.

Now that all inner modules return offsets, any fix for overflow handling at the `TimeZone` trait level will automatically be picked up without merge conflicts.