@pitdicker 2023-05-18
Split out the two commits that refactor `Local` to https://github.com/chronotope/chrono/pull/1072.