@pitdicker 2023-05-20
I have split the changes into a couple more commits, and updated them to fit with the developments on 0.4.x (`MSRV bump, `TryFrom` instead of `ToPrimitive`).