@pitdicker 2023-05-26
After the refactor in https://github.com/chronotope/chrono/pull/1072 this PR is now 20% smaller. The rebase was pretty difficult though.