@pitdicker 2023-06-23
One thing I forgot to check is the benchmark on Linux.
The logic in `lookup_with_dst_transitions` now works with `NaiveDateTime`s instead of simple Unix timestamps. Does that effect performance?

Before:
```
bench_get_local_time    time:   [91.464 ns 91.902 ns 92.373 ns]
Found 11 outliers among 100 measurements (11.00%)
  11 (11.00%) high mild
```

After:
```
bench_get_local_time    time:   [91.522 ns 91.786 ns 92.072 ns]
                        change: [-2.3806% -1.6025% -0.8457%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
```

No interesting changes.