@pitdicker 2023-06-25
During one of the rebases the test for `lookup_with_dst_transitions` disappeared. I have added it back. Sorry!