@pitdicker 2023-06-25
I have found a way to make `lookup_with_dst_transitions` *much* simpler and able to work with an arbitrary number of transitions per year.
This would be needed to fix problem 1 in https://github.com/chronotope/chrono/issues/1153 (for a later PR), and happens to fix problem 3 in https://github.com/chronotope/chrono/issues/1153 by sorting the transition dates in UTC.

@djc The change is attached as a separate commit. Do you want me to squash it into the right commits here, or remove it and open a followup PR?