@djc 2023-06-26
> @djc The change is attached as a separate commit. Do you want me to squash it into the right commits here, or remove it and open a followup PR?

How orthogonal is it to the other changes in this PR? Maybe it should go in a separate PR to land before this one?