@pitdicker 2023-06-26
If you have not reviewed `lookup_with_dst_transitions` yet, these changes can make it easier. But otherwise I am fine with moving it to a follow-up PR.