@pitdicker 2023-08-12
@djc I am not really happy with this PR. The Windows implementation that this started with can be quite simple and be written in 200 lines. And it adds an equal number of lines for tests.

But the Unix `tz_info` module is not all that integrated in chrono. It has its own types and logic that partly duplicate what chrono has. Changing it just enough so that we can have a shared lookup function for Windows and Unix makes this PR more difficult and leaves the Unix module in a state that does not feel better.

If I remove everything except the changes needed for Windows, would that help with review?
And we can refactor `tz_info` later?