@ChrisDenton 2024-01-25
Hi @pitdicker, would you mind if I helped to move this PR along? I can throw up some commits addressing @djc's review comments which you can take or leave however you like?